[build] [sourcemap] Add external sourcemap for easier production debugging.
