+++ .gitignore
@@ -191,3 +191,6 @@ $RECYCLE.BIN/
 
 # Windows shortcuts
 *.lnk
+
+# Cypres screenshots and videos
+cypress/screenshots
