+++ cypress/pages/global/index.html
@@ -8,7 +8,7 @@
         />
         <meta http-equiv="X-UA-Compatible" content="ie=edge" />
         <title>E2E Test for the Checkout Adapter Library</title>
-        <script src="/lib/global/checkout.js" async defer></script>
+        <script src="/lib/global/checkout.global.js" async defer></script>
         <script src="main.ts" type="module"></script>
     </head>
     <body>
