+++ tests/lib.adapter.test.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-require('../lib/adapter/checkout.js');
+require('../lib/adapter/checkout.adapter.js');
 
 describe('lib.adapter', () => {
     test('exposes Checkout', () => {
