+++ tests/lib.global.test.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-require('../lib/global/checkout.js');
+require('../lib/global/checkout.global.js');
 
 describe('lib.global', () => {
     test('exposes Checkout', () => {
