+++ tsconfig.json
@@ -4,7 +4,7 @@
         "useDefineForClassFields": true,
         "module": "ESNext",
         "lib": ["ESNext", "DOM"],
-        "moduleResolution": "Node",
+        "moduleResolution": "Bundler",
         "strict": true,
         "sourceMap": true,
         "resolveJsonModule": true,
